project/backend/requirements_minimal.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
python-decouple==3.8
requests==2.31.0
django-filter==23.4
django-extensions==3.2.3
